non-recursive type decs and var decs
This is just a start to what we need to accomplish, but we are able to
now add to the type and variable environments. We still need to add
function decs as well as implement recursive fun and type decs(but that
should be the last thing we do although we should think about it early).