update var and funentrys to new definition
This is not yet complete. We still need to update semant.sml to
accommadate this change. Also, 15 tests start failing because of the
access of varentry. It appears all the errors are coming when we try to
allocate a new local when we are in the top level, which is a big no
no...