add string IR function
This converts a string to IR. According to the book this is how we
convert a string to IR, but it also references a change to the
```Frame``` module, but I have no idea wtf it is asking us to do...

My best guess is that our Frame will need to be able to take our frags
and convert it to assembly. But in terms of IR, we only need the label
and to remember the value for later.